separation into components and pages. added cleaning of inputs after completing task counting
